prevent having to reload things every time...

+++ deps.sh
@@ -1,5 +1,13 @@
 #!/bin/bash
 #set -ev
+if [ -z ${DEPS_DIR} ]; then
+  echo "DEPS_DIR is not set"
+  return 1
+fi
+if [ -z ${LOCAL_DIR} ]; then
+  echo "LOCAL_DIR is not set"
+  return 1
+fi
 
 function vGood()
 {
@@ -35,11 +43,14 @@ fi
 vstring=$(bison -V | grep bison)
 if [ 0 -eq $(vGood 3.4.1 ${vstring##* }) ]; then
   echo "Bison ${vstring##* }  too low"
-  cd ${DEPS_DIR}
-  rm -rf ${DEPS_DIR}/bison
-  mkdir -p ${DEPS_DIR}/bison
-  BISON_URL="http://ftp.gnu.org/gnu/bison/bison-3.4.1.tar.gz"
-  $retry wget --no-check-certificate --quiet -O - ${BISON_URL} | tar --strip-components=1 -xz -C bison
+  if [ ! -f ${DEPS_DIR}/bison/src/bison ]; then
+    echo "Bison dep not found."
+    rm -rf ${DEPS_DIR}/bison
+    mkdir -p ${DEPS_DIR}/bison
+    cd ${DEPS_DIR}
+    BISON_URL="http://ftp.gnu.org/gnu/bison/bison-3.4.1.tar.gz"
+    $retry wget --no-check-certificate --quiet -O - ${BISON_URL} | tar --strip-components=1 -xz -C bison
+  fi
   cd ${DEPS_DIR}/bison
   ./configure --prefix=${LOCAL_DIR} --quiet
   sudo make install
@@ -48,21 +59,28 @@ fi
 vstring=$(flex -V)
 if [ 0 -eq $(vGood 2.6.3 ${vstring##* }) ]; then
   echo "Flex ${vstring##* }  too low"
-  cd ${DEPS_DIR}
-  rm -rf ${DEPS_DIR}/flex
-  mkdir -p ${DEPS_DIR}/flex
-  FLEX_URL="https://github.com/westes/flex/releases/download/v2.6.3/flex-2.6.3.tar.gz"
-  $retry wget --no-check-certificate --quiet -O - ${FLEX_URL} | tar --strip-components=1 -xz -C flex
+  if [ ! -f ${DEPS_DIR}/flex/src/flex ]; then
+    echo "Flex dep not found."
+    rm -rf ${DEPS_DIR}/flex
+    mkdir -p ${DEPS_DIR}/flex
+    cd ${DEPS_DIR}
+    FLEX_URL="https://github.com/westes/flex/releases/download/v2.6.3/flex-2.6.3.tar.gz"
+    $retry wget --no-check-certificate --quiet -O - ${FLEX_URL} | tar --strip-components=1 -xz -C flex
+  fi
   cd ${DEPS_DIR}/flex
-  ./configure --prefix=${LOCAL_DIR} --quiet && sudo make install
+  ./configure --prefix=${LOCAL_DIR} --quiet
+  sudo make install
 fi
 
 vstring=$(cmake --version | grep version)
 if [ 0 -eq $(vGood 3.13.0 ${vstring##* }) ]; then
   echo "Cmake ${vstring##* } too low"
-  cd ${DEPS_DIR}
-  rm -rf ${DEPS_DIR}/cmake
-  mkdir -p ${DEPS_DIR}/cmake
-  CMAKE_URL="https://cmake.org/files/v3.13/cmake-3.13.0-rc3-Linux-x86_64.tar.gz"
-  $retry wget --no-check-certificate --quiet -O - ${CMAKE_URL} | tar --strip-components=1 -xz -C cmake
+  if [ ! -f ${DEPS_DIR}/cmake/bin/cmake ]; then
+    echo "Cmake dep not found."
+    cd ${DEPS_DIR}
+    rm -rf ${DEPS_DIR}/cmake
+    mkdir -p ${DEPS_DIR}/cmake
+    CMAKE_URL="https://cmake.org/files/v3.13/cmake-3.13.0-rc3-Linux-x86_64.tar.gz"
+    $retry wget --no-check-certificate --quiet -O - ${CMAKE_URL} | tar --strip-components=1 -xz -C cmake
+  fi
 fi
